requirements.txt
aniso8601==1.2.1
click==6.7
docopt==0.6.2
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-RESTful==0.3.6
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
greenlet==0.4.12
inflect==0.2.5
itsdangerous==0.24
Jinja2==2.9.6
logmatic-python==0.1.7
MarkupSafe==1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.1
nose2==0.6.5
pkg-resources==0.0.0
psycopg2==2.7.1
python-dateutil==2.6.1
python-engineio==1.7.0
python-json-logger==0.1.8
python-socketio==1.8.0
pytz==2017.2
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
yamlordereddictloader==0.4.0
